feat(LottoResult): 결과를 전달할 dto 생성
